Move batch suggestion exception class
Relocates `BatchSuggestionValidationException` into `App\Exceptions` so validation failures follow the app's exception conventions, and updates the assistance controller/service imports accordingly. The tests were also adjusted to match the current nested `suggestions` payload returned for size/format assistant results.
